myproject/posts/serializers.py
from . import models
from rest_framework import serializers

class PostSerializer(serializers.ModelSerializer):

    class Meta:
        fields = ('id', 'title', 'content', 'created_at', 'updated_at',)
        model = models.Post
