myproject/requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
certifi==2018.1.18
chardet==3.0.4
cryptography==2.1.4
dj-database-url==0.5.0
Django==2.1.5
django-contrib-comments==1.9.0
django-livereload-server==0.3.2
djangorestframework==3.9.1
gunicorn==19.9.0
httplib2==0.9.2
oauth==1.0.1
Pillow==5.1.0
psycopg2==2.7.7
python-debian==0.1.32
pytz==2018.9
PyYAML==3.12
requests==2.18.4
simplejson==3.13.2
whitenoise==4.1.2
django-import-export
psycopg2-binary==2.7.7
django-cors-headers
